backend/requirements.txt
functions_framework
google-cloud-bigquery